requirements.txt
appnope
asttokens
attrs
avro
backcall
beautifulsoup4
bleach
cachetools
certifi
charset-normalizer
click
comm
contourpy
cycler
db-dtypes
debugpy
decorator
defusedxml
dill
docopt
executing
fastavro
fastjsonschema
fitter
fonttools
google-api-core
google-auth
google-cloud-bigquery
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos
idna
importlib-metadata
importlib-resources
ipykernel
ipynb
ipython
jedi
Jinja2
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_pygments
kiwisolver
loguru
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
mysql-connector-python
natsort
nbclient
nbconvert
nbformat
nest-asyncio
numpy
openturns
packaging
pandas
pandavro
pandocfilters
parso
patsy
pexpect
pg8000
pickleshare
Pillow
pipreqs
pipreqsnb
platformdirs
plotly
prompt-toolkit
protobuf
psutil
psycopg
psycopg-binary
psycopg-pool
psycopg2-binary
ptyprocess
pure-eval
pyarrow
pyasn1
pyasn1-modules
Pygments
PyMySQL
pyparsing
python-dateutil
pytz
pyzmq
referencing
reqs
requests
rich
rich-click
rpds-py
rsa
scikit-learn
scipy
seaborn
semantic-version
semver
six
soupsieve
SQLAlchemy
stack-data
statsmodels
tenacity
threadpoolctl
tinycss2
tornado
tqdm
traitlets
typing_extensions
tzdata
urllib3
wcwidth
webencodings
yarg
zipp
